Fixed code to be compliant with class standards
* implemented "Coral_X" naming convention for tube worms
+ added a new class specifically for generating tube worms to separate it from the general hydrothermic spawning logic
* moved tube worm spawning logic to the appropriate class
+ added a new hydrothermic biom generator game object instead of adding additional game objects to VoxelUniverse
