+++ Assets/Scenes/SampleScene.unity
@@ -504,6 +504,80 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1015166725}
   m_CullTransparentMesh: 0
+--- !u!1 &1190363449
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 1190363451}
+  - component: {fileID: 1190363450}
+  - component: {fileID: 1190363452}
+  m_Layer: 0
+  m_Name: HydrothermicSpawner
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &1190363450
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1190363449}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 905b65d0ec6681b40b21dbeb9485eb0f, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  tubeWormSpawner: {fileID: 1190363452}
+--- !u!4 &1190363451
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1190363449}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 3
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1190363452
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1190363449}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: a43fe3856bb158345a4ff83c30c4677b, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  mat: {fileID: 2100000, guid: a8db78d0a2516de4c84c49c9ae53d6ba, type: 2}
+  baseIterations: 14
+  baseScale: 5
+  scaleRate: 0.9
+  tubeWormScale: 0.25
+  tubeWormSpawnRate: 0.25
+  tubeWormSpawnDelay: 0.5
+  tubeWormBranchChance: 0.05
+  tubeWormSegmentLength: 2
+  wormBonusIter: 5
+  seed: 0
+  rockPrefab: {fileID: 2881934186644748434, guid: e6a0c3050443a3740a2c4e6b2b3aa8ea,
+    type: 3}
+  wormPrefab: {fileID: 33468178300371448, guid: 5c327da19862f61408f7b4302ef73176,
+    type: 3}
+  tubePrefab: {fileID: 5298450886448364052, guid: 4c882e814ab4c9d4f85a80b685cbb6f1,
+    type: 3}
 --- !u!1 &1606456750
 GameObject:
   m_ObjectHideFlags: 0
@@ -789,7 +863,6 @@ GameObject:
   m_Component:
   - component: {fileID: 2135429414}
   - component: {fileID: 2135429413}
-  - component: {fileID: 2135429415}
   m_Layer: 0
   m_Name: VoxelUniverse
   m_TagString: Untagged
@@ -846,32 +919,3 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &2135429415
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2135429412}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 905b65d0ec6681b40b21dbeb9485eb0f, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  mat: {fileID: 2100000, guid: a8db78d0a2516de4c84c49c9ae53d6ba, type: 2}
-  baseIterations: 14
-  baseScale: 5
-  scaleRate: 0.9
-  tubeWormScale: 0.25
-  tubeWormSpawnRate: 0.25
-  tubeWormSpawnDelay: 0.5
-  tubeWormBranchChance: 0.05
-  tubeWormSegmentLength: 2
-  wormBonusIter: 5
-  seed: 0
-  rockPrefab: {fileID: 2881934186644748434, guid: e6a0c3050443a3740a2c4e6b2b3aa8ea,
-    type: 3}
-  wormPrefab: {fileID: 33468178300371448, guid: 5c327da19862f61408f7b4302ef73176,
-    type: 3}
-  tubePrefab: {fileID: 5298450886448364052, guid: 4c882e814ab4c9d4f85a80b685cbb6f1,
-    type: 3}
