+++ Assets/Scripts/Coral_TubeWormSpawner.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a43fe3856bb158345a4ff83c30c4677b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
